Revert "optionally only show selected geometry without recreating scene on toggle"
This reverts commit 576e0678467d53de670f5342bd27f116886fe277.
