+++ app/pages/alpha/scene.tsx
@@ -4,7 +4,6 @@ import Head from 'next/head'
 import { Grasshopper } from 'glib'
 import { Layout, Scene } from '@/components'
 import { GraphManager } from '~/context/graph'
-import { SceneManager } from '@/components/scene/lib/context'
 import { ApolloClient, InMemoryCache } from '@apollo/client'
 import { COMPUTE_CONFIGURATION } from '@/queries'
 
@@ -20,9 +19,7 @@ const AlphaScenePage: NextPage<ScenePageProps> = ({ config }) => {
         <meta name="description" content="The scene viewer for NodePen. View the 3D results of the current graph." />
       </Head>
       <GraphManager config={config}>
-        <SceneManager>
-          <Scene.Container />
-        </SceneManager>
+        <Scene.Container />
       </GraphManager>
     </Layout.Root>
   )
